// src/libs/enums/product.enum.ts
export enum ProductStatus {
	ACTIVE = 'ACTIVE',
	FEATURED = 'FEATURED',
	SALE = 'SALE',
	PAUSED = 'PAUSED',
	DELETE = 'DELETED',
}

export enum ProductGender {
	MEN = 'MEN',
	WOMEN = 'WOMEN',
	UNISEX = 'UNISEX',
}

export enum ProductCategory {
	TOPS = 'TOPS',
	JEANS = 'JEANS',
	JACKETS = 'JACKETS',
	SWEATERS = 'SWEATERS',
	ACTIVEWEAR = 'ACTIVEWEAR',
	SHOES = 'SHOES',
	ACCESSORIES = 'ACCESSORIES',
	FORMALWEAR = 'FORMALWEAR',
	SLEEPWEAR = 'SLEEPWEAR',
	TRADITIONAL = 'TRADITIONAL',
	OTHER = 'OTHER',
}